backend/app/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LLM Providers
openai==1.3.0
boto3==1.34.0
botocore==1.34.0

# Vector Database
pinecone-client==2.2.4

# HTTP client for external APIs
httpx==0.25.2
aiohttp==3.9.1

# Data processing and utilities
numpy==1.24.3
pandas==2.1.4
python-json-logger==2.0.7

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# CORS support
fastapi-cors==0.0.6

# Optional: Database support
sqlalchemy==2.0.23
alembic==1.13.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
